requirements.txt
# -----------------------------
# Core CLI
# -----------------------------
typer==0.12.3
rich==13.9.4
click==8.1.7
shellingham==1.5.4
typing-extensions==4.12.2

# -----------------------------
# System & Utilities
# -----------------------------
psutil==5.9.8
Pillow==10.4.0
customtkinter==5.2.2
tkinterdnd2==0.4.2

# -----------------------------
# Binary Build
# -----------------------------
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.8
altgraph==0.17.4

# -----------------------------
# Testing
# -----------------------------
pytest==8.3.2
pytest-mock==3.14.0

# -----------------------------
# Linting / Formatting
# -----------------------------
ruff==0.6.9
packaging==24.1

# -----------------------------
# Markdown / Docs (optional)
# -----------------------------
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.18.0
